resume_backend/requirements.txt
accelerate==1.12.0
altair==6.0.0
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.11.1
attrs==25.4.0
blinker==1.9.0
blis==1.3.0
cachetools==6.2.2
catalogue==2.0.10
certifi==2025.11.12
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.0
cloudpathlib==0.23.0
colorama==0.4.6
confection==0.1.5
cryptography==46.0.3
cymem==2.0.11
distro==1.9.0
Django==6.0.3
django-cors-headers==4.0.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
docstring_parser==0.17.0
docx2txt==0.9
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
filelock==3.20.0
fsspec==2025.10.0
gitdb==4.0.12
GitPython==3.1.46
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-cloud-aiplatform==1.128.0
google-cloud-bigquery==3.38.0
google-cloud-core==2.5.0
google-cloud-resource-manager==1.15.0
google-cloud-storage==3.6.0
google-crc32c==1.7.1
google-genai==1.52.0
google-generativeai==0.8.5
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
gpt4all==2.8.2
greenlet==3.3.1
grpc-google-iam-v1==0.14.3
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
imaplib2==3.6
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
lxml==6.0.2
Markdown==3.10.2
MarkupSafe==3.0.3
mpmath==1.3.0
murmurhash==1.0.13
mysql-connector-python==9.5.0
narwhals==2.16.0
networkx==3.5
nltk==3.9.2
numpy==2.3.4
openai==2.8.0
packaging==25.0
pandas==2.3.3
pdfminer.six==20251107
pdfplumber==0.11.8
pgvector==0.4.2
phonenumbers==9.0.18
pillow==12.0.0
playwright==1.58.0
preshed==3.0.10
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.2.2
psycopg2-binary==2.9.11
pyarrow==23.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
pyee==13.0.1
PyJWT==2.11.0
PyMuPDF==1.26.6
pyparsing==3.2.5
PyPDF2==3.0.1
pypdfium2==5.0.0
pyresparser==1.0.6
pyrsistent==0.20.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-docx==1.2.0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
regex==2025.11.3
reportlab==4.4.4
requests==2.32.5
resume-parser==0.8.4
rpds-py==0.28.0
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.8.0
scipy==1.16.3
sentence-transformers==5.2.0
setuptools==80.9.0
shapely==2.1.2
six==1.17.0
smart_open==7.4.4
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
spacy==3.8.8
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.5
srsly==2.5.1
stemming==1.0.1
sympy==1.14.0
tenacity==9.1.2
thinc==8.3.8
threadpoolctl==3.6.0
tika==3.1.0
tokenizers==0.22.1
toml==0.10.2
torch==2.9.1
tornado==6.5.4
tqdm==4.67.1
transformers==4.57.1
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
wasabi==1.1.3
watchdog==6.0.0
weasel==0.4.2
websockets==15.0.1
wrapt==2.0.1
